# sca-web/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/subject/{id}":{"get":{"tags":["subject-controller"],"operationId":"getSubjectById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SubjectResponseDto"}}}}}},"put":{"tags":["subject-controller"],"operationId":"updateSubject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subject"}}}}}},"delete":{"tags":["subject-controller"],"operationId":"deleteSubjectById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/studyClass/{id}":{"get":{"tags":["study-class-controller"],"operationId":"getStudyClassById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudyClass"}}}}}},"put":{"tags":["study-class-controller"],"operationId":"updateStudyClass","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudyClassRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["study-class-controller"],"operationId":"deleteStudyClassById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/student/{id}":{"get":{"tags":["user-controller"],"operationId":"getStudentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudentResponseDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateStudent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteStudentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/professor/{id}":{"get":{"tags":["user-controller"],"operationId":"getProfessorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfessorResponseDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateProfessor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfessorRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Professor"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteProfessorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/subject":{"get":{"tags":["subject-controller"],"operationId":"fetchSubjectList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Subject"}}}}}}},"post":{"tags":["subject-controller"],"summary":"createSubject","description":"Create subject","operationId":"saveSubject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectRequestDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subject"}}}}}}},"/studyClass":{"get":{"tags":["study-class-controller"],"operationId":"fetchStudyClassList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudyClass"}}}}}}},"post":{"tags":["study-class-controller"],"summary":"createStudyClass","description":"Create studyClass","operationId":"saveStudyClass","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudyClassRequestDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StudyClass"}}}}}}},"/student":{"get":{"tags":["user-controller"],"operationId":"fetchStudentList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentDto"}}}}}}},"post":{"tags":["user-controller"],"summary":"createStudent","description":"Create student","operationId":"saveStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRequestDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}}},"/professor":{"get":{"tags":["user-controller"],"operationId":"fetchProfessorList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProfessorDto"}}}}}}},"post":{"tags":["user-controller"],"operationId":"saveProfessor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfessorRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Professor"}}}}}}}},"components":{"schemas":{"SubjectRequestDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"workload":{"type":"integer","format":"int32"},"creditNumber":{"type":"integer","format":"int32"}}},"Address":{"type":"object","properties":{"street":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"postalCode":{"type":"string"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"Professor":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"curriculum":{"type":"string"},"title":{"type":"string","enum":["PROFESSOR","ASSOCIATE_PROFESSOR","ASSISTANT_PROFESSOR","INSTRUCTOR","LECTURER"]},"contractType":{"type":"string","enum":["FULL_TIME","PART_TIME","VISITING","ADJUNCT"]},"studyClasses":{"type":"array","items":{"$ref":"#/components/schemas/StudyClass"}},"address":{"$ref":"#/components/schemas/Address"}}},"Student":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"gpt":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"},"studyClasses":{"type":"array","items":{"$ref":"#/components/schemas/StudyClass"}}}},"StudyClass":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"code":{"type":"string"},"name":{"type":"string"},"classroom":{"type":"string"},"description":{"type":"string"},"startTime":{"$ref":"#/components/schemas/LocalTime"},"subject":{"$ref":"#/components/schemas/Subject"},"professor":{"$ref":"#/components/schemas/Professor"},"students":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}},"Subject":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"workload":{"type":"integer","format":"int32"},"creditNumber":{"type":"integer","format":"int32"},"studyClasses":{"type":"array","items":{"$ref":"#/components/schemas/StudyClass"}}}},"StudyClassRequestDto":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"classroom":{"type":"string"},"description":{"type":"string"},"startTime":{"type":"string"},"professorId":{"type":"integer","format":"int32"},"subjectId":{"type":"integer","format":"int32"},"studentsIds":{"type":"array","items":{"type":"integer","format":"int32"}}}},"ProfessorRequestDto":{"type":"object","properties":{"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"curriculum":{"type":"string"},"title":{"type":"string","enum":["PROFESSOR","ASSOCIATE_PROFESSOR","ASSISTANT_PROFESSOR","INSTRUCTOR","LECTURER"]},"contractType":{"type":"string","enum":["FULL_TIME","PART_TIME","VISITING","ADJUNCT"]},"address":{"$ref":"#/components/schemas/Address"}}},"StudentRequestDto":{"type":"object","properties":{"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"gpt":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"}}},"StudyClassDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"code":{"type":"string"},"name":{"type":"string"},"classroom":{"type":"string"},"description":{"type":"string"},"startTime":{"type":"string"}}},"SubjectResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"workload":{"type":"integer","format":"int32"},"creditNumber":{"type":"integer","format":"int32"},"studyClasses":{"type":"array","items":{"$ref":"#/components/schemas/StudyClassDto"}}}},"StudentDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"gpt":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"}}},"StudentResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"gpt":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"},"studyClasses":{"type":"array","items":{"$ref":"#/components/schemas/StudyClassDto"}}}},"ProfessorDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"curriculum":{"type":"string"},"title":{"type":"string","enum":["PROFESSOR","ASSOCIATE_PROFESSOR","ASSISTANT_PROFESSOR","INSTRUCTOR","LECTURER"]},"contractType":{"type":"string","enum":["FULL_TIME","PART_TIME","VISITING","ADJUNCT"]},"address":{"$ref":"#/components/schemas/Address"}}},"ProfessorResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"registration":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"curriculum":{"type":"string"},"title":{"type":"string","enum":["PROFESSOR","ASSOCIATE_PROFESSOR","ASSISTANT_PROFESSOR","INSTRUCTOR","LECTURER"]},"contractType":{"type":"string","enum":["FULL_TIME","PART_TIME","VISITING","ADJUNCT"]},"address":{"$ref":"#/components/schemas/Address"},"studyClasses":{"type":"array","items":{"$ref":"#/components/schemas/StudyClassDto"}}}}}}}